Write a function that returns the sum of all its parameters.
    Prototype: int sum_them_all(const unsigned int n, ...);
    If n == 0, return 0

+++ variadic_functions/0-sum_them_all.c
@@ -3,20 +3,19 @@
 /**
  * sum_them_all - function that returns the sum of all its parameters
  * @n: number of args
- * @...: variables number
  * Return: sum (or 0 if n == 0)
 */
 int sum_them_all(const unsigned int n, ...)
 {
 	va_list args;
 	unsigned int i;
-	double sum = 0;
+	int sum = 0;
 
 	if (n == 0)
 		return (0);
 	va_start(args, n);
 	for (i = 0; i < n; i++)
-		sum += va_arg(args, double);
+		sum += va_arg(args, int);
 	va_end(args);
 	return (sum);
 }
